+++ packages/components/inputs/search-select-input/src/search-select-input.styles.js
@@ -1,0 +1,19 @@
+/* eslint-disable import/prefer-default-export */
+import styled from '@emotion/styled';
+import { customProperties as vars } from '@commercetools-uikit/design-system';
+
+const SearchSelectInputWrapper = styled.div`
+  div[class$='control'] {
+    &:hover {
+      cursor: text;
+    }
+    div[class$='indicatorContainer' i] {
+      cursor: pointer;
+      svg * {
+        fill: ${vars.colorSolid};
+      }
+    }
+  }
+`;
+
+export { SearchSelectInputWrapper };