+++ packages/components/inputs/money-input/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @commercetools-uikit/money-input
 
+## 10.36.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @commercetools-uikit/select-utils@10.36.0
+
 ## 10.35.2
 
 ### Patch Changes