+++ packages/components/inputs/rich-text-input/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @commercetools-uikit/rich-text-input
 
+## 10.36.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @commercetools-uikit/flat-button@10.36.0
+
 ## 10.35.1
 
 ### Patch Changes